refactored query into a hook
